# testkite/packages/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TestKite Contract",
    "version": "0.0.1",
    "description": "Catalog schema authoring-facing của TestKite, sinh từ zod. M1: chỉ components.schemas — paths gắn ở M2 cùng route Fastify."
  },
  "paths": {
...
    "/v1/projects/{projectId}/cases": {
      "post": {
        "operationId": "createCase",
        "summary": "Create a case inside a project",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "isStepGroup": {
                    "type": "boolean",
                    "default": false
                  },
                  "prereqCaseId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseSummary"
                }
              }
            }
          },
          "403": {
            "description": "Thiếu quyền trong chính team của mình",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy — GỒM CẢ tài nguyên của team khác (không bao giờ 403)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/cases/{caseId}/steps": {
      "put": {
        "operationId": "replaceSteps",
        "summary": "Replace a case's steps (optimistic concurrency: requires If-Match)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "steps": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/StepInput"
                    }
                  }
                },
                "required": [
                  "steps"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseSummary"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Thiếu quyền trong chính team của mình",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy — GỒM CẢ tài nguyên của team khác (không bao giờ 403)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Xung đột phiên bản",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          },
          "428": {
            "description": "Thiếu If-Match",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/cases/{caseId}/withdraw-review": {
      "post": {
        "operationId": "withdrawReview",
        "summary": "Withdraw a case from review back to draft (requires If-Match)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseSummary"
                }
              }
            }
          },
          "403": {
            "description": "Thiếu quyền trong chính team của mình",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy — GỒM CẢ tài nguyên của team khác (không bao giờ 403)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Xung đột phiên bản",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          },
          "428": {
            "description": "Thiếu If-Match",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/cases/{caseId}/promote": {
      "post": {
        "operationId": "promoteCase",
        "summary": "Promote an approved case to ready (four-eyes; requires If-Match)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseSummary"
                }
              }
            }
          },
          "403": {
            "description": "Thiếu quyền trong chính team của mình",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy — GỒM CẢ tài nguyên của team khác (không bao giờ 403)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Xung đột phiên bản",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          },
          "428": {
            "description": "Thiếu If-Match",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token của TestKite: `Authorization: Bearer tk_<prefix>_<secret>`. Token gắn ĐÚNG MỘT team và luôn có hạn."
      }
    },
    "schemas": {
...
      "StepInput": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "const": "action"
              },
              "id": {
                "type": "string",
                "minLength": 1
              },
              "renderedSentence": {
                "type": "string",
                "minLength": 1
              },
              "verbOpKey": {
                "type": "string",
                "minLength": 1
              },
              "args": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              },
              "elementId": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "kind",
              "renderedSentence",
              "verbOpKey"
            ]
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "const": "step_group"
              },
              "id": {
                "type": "string",
                "minLength": 1
              },
              "renderedSentence": {
                "type": "string",
                "minLength": 1
              },
              "stepGroupCaseId": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "kind",
              "renderedSentence",
              "stepGroupCaseId"
            ]
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "const": "if"
              },
              "id": {
                "type": "string",
                "minLength": 1
              },
              "renderedSentence": {
                "type": "string",
                "minLength": 1
              },
              "conditionExpected": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 1
                },
                "minItems": 1
              },
              "children": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/StepInput"
                }
              }
            },
            "required": [
              "kind",
              "renderedSentence",
              "conditionExpected",
              "children"
            ]
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "const": "for"
              },
              "id": {
                "type": "string",
                "minLength": 1
              },
              "renderedSentence": {
                "type": "string",
                "minLength": 1
              },
              "loopDataProfileId": {
                "type": "string",
                "minLength": 1
              },
              "children": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/StepInput"
                }
              }
            },
            "required": [
              "kind",
              "renderedSentence",
              "loopDataProfileId",
              "children"
            ]
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "const": "while"
              },
              "id": {
                "type": "string",
                "minLength": 1
              },
              "renderedSentence": {
                "type": "string",
                "minLength": 1
              },
              "maxIterations": {
                "type": "integer",
                "exclusiveMinimum": 0
              },
              "children": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/StepInput"
                }
              }
            },
            "required": [
              "kind",
              "renderedSentence",
              "children"
            ]
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "const": "rest"
              },
              "id": {
                "type": "string",
                "minLength": 1
              },
              "renderedSentence": {
                "type": "string",
                "minLength": 1
              },
              "method": {
                "type": "string",
                "enum": [
                  "GET",
                  "POST",
                  "PUT",
                  "PATCH",
                  "DELETE",
                  "HEAD",
                  "OPTIONS"
                ]
              },
              "url": {
                "type": "string",
                "minLength": 1
              },
              "headers": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              },
              "body": {
                "type": "string"
              },
              "storeAs": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "kind",
              "renderedSentence",
              "method",
              "url"
            ]
          }
        ]
      },
      "CaseSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "projectId": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "isStepGroup": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "in_review",
              "ready"
            ]
          },
          "version": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "prereqCaseId": {
            "type": "string",
            "minLength": 1
          },
          "dataProfileId": {
            "type": "string",
            "minLength": 1
          },
          "latestRevisionId": {
            "type": "string",
            "minLength": 1
          },
          "readyRevisionId": {
            "type": "string",
            "minLength": 1
          },
          "lastEditedBy": {
            "type": "string",
            "minLength": 1
          },
          "createdAt": {
            "type": "string",
            "minLength": 1
          },
          "updatedAt": {
            "type": "string",
            "minLength": 1
          },
          "submittedAt": {
            "type": "string",
            "minLength": 1
          },
          "reviewedAt": {
            "type": "string",
            "minLength": 1
          },
          "promotedAt": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "isStepGroup",
          "status",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}